# swagger.yaml
openapi: 3.0.3
info:
  title: e-commerce
  version: 1.0.0
  contact: {}
servers:
  - url: http://localhost:3000
paths:
  /api/categories:
    get:
      tags:
        - categories
      summary: categories
      description: categories
      operationId: categories
      parameters:
        - name: ''
          in: query
          schema:
            type: string
            example: ''
      responses:
        '200':
          description: success / after deletion
          headers:
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '191'
            Date:
              schema:
                type: string
                example: Thu, 07 Nov 2024 13:44:37 GMT
            ETag:
              schema:
                type: string
                example: W/"bf-K1RNWMft44DMZZ0V6U0geWKBCQo"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        __v:
                          type: number
                          example: 0
                        _id:
                          type: string
                          example: 672cc3acbb8bb89e108ce5ec
                        createdAt:
                          type: string
                          example: '2024-11-07T13:42:04.546Z'
                        name:
                          type: string
                          example: Fashion
                        updatedAt:
                          type: string
                          example: '2024-11-07T13:42:04.546Z'
                    example:
                      - __v: 0
                        _id: 672cc3acbb8bb89e108ce5ec
                        createdAt: '2024-11-07T13:42:04.546Z'
                        name: Fashion
                        updatedAt: '2024-11-07T13:42:04.546Z'
                      - __v: 0
                        _id: 672cc3c9bb8bb89e108ce5ee
                        createdAt: '2024-11-07T13:42:33.103Z'
                        name: Electronics
                        updatedAt: '2024-11-07T13:42:33.103Z'
                  message:
                    type: string
                    example: Success get all categories
              examples:
                after deletion:
                  value:
                    data:
                      - __v: 0
                        _id: 672cc3c9bb8bb89e108ce5ee
                        createdAt: '2024-11-07T13:42:33.103Z'
                        name: Electronics
                        updatedAt: '2024-11-07T13:42:33.103Z'
                    message: Success get all categories
                success:
                  value:
                    data:
                      - __v: 0
                        _id: 672cc3acbb8bb89e108ce5ec
                        createdAt: '2024-11-07T13:42:04.546Z'
                        name: Fashion
                        updatedAt: '2024-11-07T13:42:04.546Z'
                      - __v: 0
                        _id: 672cc3c9bb8bb89e108ce5ee
                        createdAt: '2024-11-07T13:42:33.103Z'
                        name: Electronics
                        updatedAt: '2024-11-07T13:42:33.103Z'
                    message: Success get all categories
    post:
      tags:
        - categories
      summary: category
      description: category
      operationId: category
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Fashion
            examples:
              category:
                value:
                  name: Fashion
      responses:
        '201':
          description: success
          headers:
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '182'
            Date:
              schema:
                type: string
                example: Thu, 07 Nov 2024 13:42:04 GMT
            ETag:
              schema:
                type: string
                example: W/"b6-wVy43ryOGabbdfYM+aW21LJ8Dyo"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      __v:
                        type: number
                        example: 0
                      _id:
                        type: string
                        example: 672cc3acbb8bb89e108ce5ec
                      createdAt:
                        type: string
                        example: '2024-11-07T13:42:04.546Z'
                      name:
                        type: string
                        example: Fashion
                      updatedAt:
                        type: string
                        example: '2024-11-07T13:42:04.546Z'
                  message:
                    type: string
                    example: Success create category
              examples:
                success:
                  value:
                    data:
                      __v: 0
                      _id: 672cc3acbb8bb89e108ce5ec
                      createdAt: '2024-11-07T13:42:04.546Z'
                      name: Fashion
                      updatedAt: '2024-11-07T13:42:04.546Z'
                    message: Success create category
      security:
        - bearerAuth: []
...
  /api/products:
    get:
      tags:
        - products
      summary: products
      description: products
      operationId: products
      parameters:
        - name: page
          in: query
          schema:
            type: string
            example: '1'
        - name: limit
          in: query
          schema:
            type: string
            example: '1'
      responses:
        '200':
          description: success / after update
          headers:
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '762'
            Date:
              schema:
                type: string
                example: Fri, 08 Nov 2024 14:23:08 GMT
            ETag:
              schema:
                type: string
                example: W/"2fa-4adFe6uYKMhJg9qQLre1+b7NSD0"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        __v:
                          type: number
                          example: 0
                        _id:
                          type: string
                          example: 672e1d8d67f68c1a07655334
                        category:
                          type: object
                          properties:
                            __v:
                              type: number
                              example: 0
                            _id:
                              type: string
                              example: 672cc50ebb8bb89e108ce5f6
                            createdAt:
                              type: string
                              example: '2024-11-07T13:47:58.219Z'
                            name:
                              type: string
                              example: Fashion Female
                            updatedAt:
                              type: string
                              example: '2024-11-07T13:49:06.393Z'
                        createdAt:
                          type: string
                          example: '2024-11-08T14:17:49.038Z'
                        description:
                          type: string
                          example: daster
                        images:
                          type: array
                          items: {}
                          example: []
                        name:
                          type: string
                          example: Abaya
                        price:
                          type: number
                          example: 100000
                        qty:
                          type: number
                          example: 10
                        updatedAt:
                          type: string
                          example: '2024-11-08T14:17:49.038Z'
                    example:
                      - __v: 0
                        _id: 672e1d8d67f68c1a07655334
                        category:
                          __v: 0
                          _id: 672cc50ebb8bb89e108ce5f6
                          createdAt: '2024-11-07T13:47:58.219Z'
                          name: Fashion Female
                          updatedAt: '2024-11-07T13:49:06.393Z'
                        createdAt: '2024-11-08T14:17:49.038Z'
                        description: daster
                        images: []
                        name: Abaya
                        price: 100000
                        qty: 10
                        updatedAt: '2024-11-08T14:17:49.038Z'
                      - __v: 0
                        _id: 672e1e5758d68660a938d21d
                        category:
                          __v: 0
                          _id: 672cc50ebb8bb89e108ce5f6
                          createdAt: '2024-11-07T13:47:58.219Z'
                          name: Fashion Female
                          updatedAt: '2024-11-07T13:49:06.393Z'
                        createdAt: '2024-11-08T14:21:11.860Z'
                        description: daster
                        images: []
                        name: Abaya Gamis
                        price: 100000
                        qty: 10
                        updatedAt: '2024-11-08T14:21:11.860Z'
                  message:
                    type: string
                    example: Success get all products
              examples:
                after update:
                  value:
                    data:
                      - __v: 0
                        _id: 672e1d8d67f68c1a07655334
                        category:
                          __v: 0
                          _id: 672cc50ebb8bb89e108ce5f6
                          createdAt: '2024-11-07T13:47:58.219Z'
                          name: Fashion Female
                          updatedAt: '2024-11-07T13:49:06.393Z'
                        createdAt: '2024-11-08T14:17:49.038Z'
                        description: daster
                        images: []
                        name: Abaya
                        price: 100000
                        qty: 10
                        updatedAt: '2024-11-08T14:17:49.038Z'
                      - __v: 0
                        _id: 672e1e5758d68660a938d21d
                        category:
                          __v: 0
                          _id: 672cc50ebb8bb89e108ce5f6
                          createdAt: '2024-11-07T13:47:58.219Z'
                          name: Fashion Female
                          updatedAt: '2024-11-07T13:49:06.393Z'
                        createdAt: '2024-11-08T14:21:11.860Z'
                        description: daster
                        images: []
                        name: Abaya Gamis Import
                        price: 100000
                        qty: 10
                        updatedAt: '2024-11-08T14:22:57.599Z'
                    message: Success get all products
                success:
                  value:
                    data:
                      - __v: 0
                        _id: 672e1d8d67f68c1a07655334
                        category:
                          __v: 0
                          _id: 672cc50ebb8bb89e108ce5f6
                          createdAt: '2024-11-07T13:47:58.219Z'
                          name: Fashion Female
                          updatedAt: '2024-11-07T13:49:06.393Z'
                        createdAt: '2024-11-08T14:17:49.038Z'
                        description: daster
                        images: []
                        name: Abaya
                        price: 100000
                        qty: 10
                        updatedAt: '2024-11-08T14:17:49.038Z'
                      - __v: 0
                        _id: 672e1e5758d68660a938d21d
                        category:
                          __v: 0
                          _id: 672cc50ebb8bb89e108ce5f6
                          createdAt: '2024-11-07T13:47:58.219Z'
                          name: Fashion Female
                          updatedAt: '2024-11-07T13:49:06.393Z'
                        createdAt: '2024-11-08T14:21:11.860Z'
                        description: daster
                        images: []
                        name: Abaya Gamis
                        price: 100000
                        qty: 10
                        updatedAt: '2024-11-08T14:21:11.860Z'
                    message: Success get all products
    post:
      tags:
        - products
      summary: product
      description: product
      operationId: product
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: Deskripsi kemeja variant 3
                name:
                  type: string
                  example: Kemeja Variant 3
                price:
                  type: number
                  example: 1000
                qty:
                  type: number
                  example: 100
                slug:
                  type: string
                  example: variant-3-slug
            examples:
              product:
                value:
                  description: Deskripsi kemeja variant 3
                  name: Kemeja Variant 3
                  price: 1000
                  qty: 100
                  slug: variant-3-slug
      responses:
        '201':
          description: success
          headers:
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '182'
            Date:
              schema:
                type: string
                example: Thu, 07 Nov 2024 13:42:04 GMT
            ETag:
              schema:
                type: string
                example: W/"b6-wVy43ryOGabbdfYM+aW21LJ8Dyo"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      __v:
                        type: number
                        example: 0
                      _id:
                        type: string
                        example: 673dbbac5b100dc6a2efc25e
                      createdAt:
                        type: string
                        example: '2024-11-20T10:36:28.955Z'
                      description:
                        type: string
                        example: Deskripsi kemeja variant 3
                      images:
                        type: array
                        items: {}
                        example: []
                      name:
                        type: string
                        example: Kemeja Variant 3
                      price:
                        type: number
                        example: 1000
                      qty:
                        type: number
                        example: 100
                      updatedAt:
                        type: string
                        example: '2024-11-20T10:36:28.955Z'
                  message:
                    type: string
                    example: Success create product
              examples:
                success:
                  value:
                    data:
                      __v: 0
                      _id: 673dbbac5b100dc6a2efc25e
                      createdAt: '2024-11-20T10:36:28.955Z'
                      description: Deskripsi kemeja variant 3
                      images: []
                      name: Kemeja Variant 3
                      price: 1000
                      qty: 100
                      updatedAt: '2024-11-20T10:36:28.955Z'
                    message: Success create product
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...